refactor: 일가 작성 api 관련 타입을 api 호출 파일에서 타입 파일로 분리
